+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # MattyLang Changelog
 
+## v0.3.0
+- Functions
+- Return Statement
+
+## v0.2.0
+- If Statement
+- While Statement
+- Break, Continue Statements
+
+## v0.1.0
+- Variables
+- Expressions
+
 ## v0.0.0
 - Created project